Conflicts:
	searchManager/WebContent/js/settings/import.js

+++ searchManager/src/com/search/manager/service/RuleTransferService.java
@@ -1,5 +1,6 @@
 package com.search.manager.service;
 
+import java.io.File;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -23,11 +24,13 @@ import com.search.manager.enums.ExportType;
 import com.search.manager.enums.ImportType;
 import com.search.manager.enums.RuleEntity;
 import com.search.manager.enums.RuleStatusEntity;
+import com.search.manager.model.AuditTrail;
 import com.search.manager.model.ExportRuleMap;
 import com.search.manager.model.FacetSort;
 import com.search.manager.model.RecordSet;
 import com.search.manager.model.RuleStatus;
 import com.search.manager.model.SearchCriteria;
+import com.search.manager.model.constants.AuditTrailConstants;
 import com.search.manager.report.model.xml.RuleXml;
 import com.search.manager.xml.file.RuleTransferUtil;
 import com.search.manager.xml.file.RuleXmlUtil;
@@ -72,6 +75,12 @@ public class RuleTransferService {
 		String store = UtilityService.getStoreName();
 		RuleEntity ruleEntity = RuleEntity.find(ruleType);
 		List<String> successList = new ArrayList<String>();
+		List<String> successRuleStatusIdList = new ArrayList<String>();
+
+		AuditTrail auditTrail = new AuditTrail();
+		auditTrail.setOperation(String.valueOf(AuditTrailConstants.Operation.exportRule));
+		auditTrail.setUsername(UtilityService.getUsername());
+		auditTrail.setStoreId(store);
 
 		for(int i = 0 ; i < ruleRefIdList.length; i++){
 			String ruleId = ruleRefIdList[i];
@@ -80,12 +89,27 @@ public class RuleTransferService {
 			if(RuleTransferUtil.exportRule(store, ruleEntity, ruleId, ruleXml)){
 				RuleStatus ruleStatus = new RuleStatus(RuleEntity.getId(ruleType), store, ruleId);
 				SearchCriteria<RuleStatus> searchCriteria =new SearchCriteria<RuleStatus>(ruleStatus,null,null,null,null);
+				
 				try {
 					RecordSet<RuleStatus> approvedRset = daoService.getRuleStatus(searchCriteria);
 					if (approvedRset.getTotalSize() > 0) {
 						ruleStatus = approvedRset.getList().get(0);
-						daoService.updateRuleStatusExportInfo(ruleStatus, UtilityService.getUsername(), ExportType.MANUAL, new Date());
-						successList.add(ruleId);
+						
+						if(ruleStatus != null){
+							auditTrail.setEntity(String.valueOf(AuditTrailConstants.Entity.ruleStatus));
+							auditTrail.setDate(new Date());
+							auditTrail.setReferenceId(ruleStatus.getRuleRefId());
+							if (ruleEntity == RuleEntity.ELEVATE || ruleEntity == RuleEntity.EXCLUDE || ruleEntity == RuleEntity.DEMOTE) {
+								auditTrail.setKeyword(ruleStatus.getRuleRefId());
+							}
+							auditTrail.setDetails(String.format("Exported reference id = [%1$s], rule type = [%2$s], export type = [%3$s].", 
+									auditTrail.getReferenceId(), RuleEntity.getValue(ruleStatus.getRuleTypeId()), ExportType.MANUAL));
+							daoService.addAuditTrail(auditTrail);
+
+							daoService.updateRuleStatusExportInfo(ruleStatus, UtilityService.getUsername(), ExportType.MANUAL, new Date());
+							successList.add(getSuccessRule(ruleEntity, ruleId, ruleStatus.getRuleName()));
+							successRuleStatusIdList.add(ruleStatus.getRuleStatusId());
+						}
 					}
 					else {
 						logger.error("No rule status found for " + ruleEntity + " : "  + ruleId);
@@ -95,16 +119,24 @@ public class RuleTransferService {
 				}
 			}
 		}
+		daoService.addRuleStatusComment(null, "[EXPORTED] " + comment, successRuleStatusIdList.toArray(new String[0]));
 		return successList;
 	}
 
 	@RemoteMethod
 	public List<String> importRules(String ruleType, String[] ruleRefIdList, String comment, String[] importTypeList, String[] importAsRefIdList, String[] ruleNameList){
 		List<String> successList = new ArrayList<String>();
+		List<String> importedRuleStatusIds = new ArrayList<String>();
 		String store = UtilityService.getStoreName();
 		RuleEntity ruleEntity = RuleEntity.find(ruleType);
 		String userName = UtilityService.getUsername();
 		
+		AuditTrail auditTrail = new AuditTrail();
+		auditTrail.setOperation(String.valueOf(AuditTrailConstants.Operation.importRule));
+		auditTrail.setUsername(UtilityService.getUsername());
+		auditTrail.setStoreId(store);
+		auditTrail.setEntity(String.valueOf(AuditTrailConstants.Entity.ruleStatus));
+
 		for(int i = 0 ; i < ruleRefIdList.length; i++){
 			String ruleId = ruleRefIdList[i];
 			ImportType importType = ImportType.get(importTypeList[i]);
@@ -144,9 +176,31 @@ public class RuleTransferService {
 					daoService.addRuleStatus(new RuleStatus(ruleEntity, store, importAsId, ruleName, userName, userName, 
 							RuleStatusEntity.ADD, RuleStatusEntity.UNPUBLISHED));
 					status++;
+					RuleStatus ruleStatus = new RuleStatus(RuleEntity.getId(ruleType), store, ruleId);
+					SearchCriteria<RuleStatus> searchCriteria = new SearchCriteria<RuleStatus>(ruleStatus,null,null,null,null);
+					try {
+						RecordSet<RuleStatus> rSet = daoService.getRuleStatus(searchCriteria);
+						if (rSet != null && CollectionUtils.isNotEmpty(rSet.getList())) {
+							importedRuleStatusIds.add(rSet.getList().get(0).getRuleStatusId());
+							auditTrail.setDate(new Date());
+							auditTrail.setReferenceId(ruleStatus.getRuleRefId());
+							if (ruleEntity == RuleEntity.ELEVATE || ruleEntity == RuleEntity.EXCLUDE || ruleEntity == RuleEntity.DEMOTE) {
+								auditTrail.setKeyword(ruleStatus.getRuleRefId());
+							}
+							auditTrail.setDetails(String.format("Imported reference id = [%1$s], rule type = [%2$s].", 
+									auditTrail.getReferenceId(), RuleEntity.getValue(ruleStatus.getRuleTypeId())));
+							daoService.addAuditTrail(auditTrail);
+						}
+						else {
+							logger.error("No rule status found for " + ruleEntity + " : "  + ruleId);
+						}
+					} catch (DaoException e) {
+						logger.error("Failed to get rule status for " + ruleEntity + " : "  + ruleId, e);
+					}
+					
 					if(ImportType.FOR_APPROVAL == importType || ImportType.AUTO_PUBLISH == importType){
 						//submit rule for approval
-						RuleStatus ruleStatus = deploymentService.processRuleStatus(ruleType, importAsId, ruleName, false);
+						ruleStatus = deploymentService.processRuleStatus(ruleType, importAsId, ruleName, false);
 						status++;
 						if(ruleStatus != null && ImportType.AUTO_PUBLISH == importType){
 							//approve rule
@@ -158,18 +212,8 @@ public class RuleTransferService {
 						}
 					}
 					
-					switch(ruleEntity){
-						case ELEVATE:
-						case EXCLUDE:
-						case DEMOTE:
-							successList.add(ruleId);
-							break;
-						case FACET_SORT:
-						case QUERY_CLEANING:
-						case RANKING_RULE:
-							successList.add(ruleName);
-							break;
-					}
+					successList.add(getSuccessRule(ruleEntity, ruleId, ruleName));
+					
 				} catch (DaoException de) {
 					String msg = "";
 					switch (status) {
@@ -183,9 +227,25 @@ public class RuleTransferService {
 				
 			}
 		}
+		daoService.addRuleStatusComment(null, "[IMPORTED] " + comment, importedRuleStatusIds.toArray(new String[0]));
 		return successList;
 	}
 
+	private String getSuccessRule(RuleEntity ruleEntity, String ruleId, String ruleName){
+		switch(ruleEntity){
+			case ELEVATE:
+			case EXCLUDE:
+			case DEMOTE:
+				return ruleId;
+			case FACET_SORT:
+			case QUERY_CLEANING:
+			case RANKING_RULE:
+				return ruleName;
+			default:
+				return ruleName;
+		}
+	}
+	
 	public boolean importRule(RuleEntity ruleEntity, String store, String ruleId, String comment, ImportType importType, String importAsRefId, String ruleName){
 		String id = RuleXmlUtil.getRuleId(ruleEntity, ruleId);
 		RuleXml ruleXml = RuleTransferUtil.getRuleToImport(store, ruleEntity, id);
@@ -220,21 +280,28 @@ public class RuleTransferService {
 		for(int i = 0 ; i < ruleRefIdList.length; i++){
 			String ruleId = ruleRefIdList[i];
 			String ruleName = ruleNameList[i];
+			String refId = ruleId;
 
-			if(deleteRuleFile(ruleEntity, store, ruleId, comment)){
+			switch(ruleEntity){
+			case ELEVATE:
+			case EXCLUDE:
+			case DEMOTE:
+				refId = ruleNameList[i]; //delete file by rule name
+				break;
+			case FACET_SORT:
+			case QUERY_CLEANING:
+			case RANKING_RULE:
+				refId = ruleRefIdList[i]; //delete file by rule id
+				break;
+			default:
+				break;
+			}
+			
+			if(deleteRuleFile(ruleEntity, store, refId, comment)){
 				//TODO addComment
 				//TODO addAuditTrail
-				switch(ruleEntity){
-				case ELEVATE:
-				case EXCLUDE:
-				case DEMOTE:
-					successList.add(ruleId);
-					break;
-				case FACET_SORT:
-				case QUERY_CLEANING:
-				case RANKING_RULE:
-					successList.add(ruleName);
-				}
+				
+				successList.add(getSuccessRule(ruleEntity, ruleId, ruleName));
 			}
 		}
 		return successList;
@@ -244,11 +311,21 @@ public class RuleTransferService {
 		boolean success = false;
 		String id = RuleXmlUtil.getRuleId(ruleEntity, ruleId);
 		try{
-			RuleXmlUtil.deleteFile(RuleTransferUtil.getFilename(store, ruleEntity, id));
-			success = true;
+			String filepath = RuleTransferUtil.getFilename(store, ruleEntity, id);
+			File file = new File(filepath);
+			
+			if(!file.exists()){
+				logger.info("File to delete not found. Filename = " + filepath);
+				success = false;
+			}
+			else{
+				RuleXmlUtil.deleteFile(RuleTransferUtil.getFilename(store, ruleEntity, id));
+				success = true;
+			}
 		}
 		catch (Exception e) {
 			e.printStackTrace();
+			return false;
 		}
 
 		return success;
